+++ c2mon-server/c2mon-server-elasticsearch/src/main/java/cern/c2mon/server/elasticsearch/config/ElasticsearchProperties.java
@@ -20,12 +20,21 @@ import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 /**
+ * Elasticsearch module properties
+ *
  * @author Justin Lewis Salmon
  * @author Serhiy Boychenko
  */
 @Data
 @ConfigurationProperties(prefix = "c2mon.server.elasticsearch")
 public class ElasticsearchProperties {
+
+  /**
+   * Type is being removed in Elasticsearch 6.x (check
+   * <a href="https://www.elastic.co/guide/en/elasticsearch/reference/master/removal-of-types.html">Elasticsearch
+   * documentation</a> for more details).
+   */
+  public static final String TYPE = "doc";
 
   /**
    * Enable/Disable writing to Elasticsearch
@@ -40,9 +49,22 @@ public class ElasticsearchProperties {
   /**
    * Port number on which to communicate
    */
-  private int port = 9300;
+  private int port = 9200;
 
-  private int httpPort = 9200;
+  /**
+   * Scheme to be used to communicate with Elasticsearch server (only application for REST client)
+   */
+  private String scheme = "http";
+
+  /**
+   * Username to be used to authenticate with Elasticsearch server
+   */
+  private String username;
+
+  /**
+   * Password to be used to authenticate with Elasticsearch server
+   */
+  private String password;
 
   /**
    * Name of the Elasticsearch cluster to connect to. Must be the same for all
@@ -142,9 +164,9 @@ public class ElasticsearchProperties {
   private String supervisionFallbackFile = "/tmp/es-supervision-fallback.txt";
 
   /**
-   * Enable/Disable Elasticsearch REST Client (if disabled Transport client will be used instead)
+   * Defines the client to be used to communicate with Elasticsearch (possible values: [rest, transport])
    */
-  private boolean rest = true;
+  private String client = "rest";
 
   /**
    * Defines whether mapping templates are managed by C2MON